Pixel art series added
I decided to start with the pixel art series to get a feel of the
workflow with the script mentioned on #18

Regarding the script, I can already tell a few things that can be
improved:
 - There's a lot of copy and paste involved, it generates the markdown
 files in the same directory of the script, you then have to manually
 copy/move everything to the correct place. It would be easier if the
 script could do it automatically, using an option when calling it
 through the cli
 - The _index.md for the chapter could be created automatically when
 creating the above
 - Same for the _index.md of the whole series

> Please adjust the date so that the courses are more or less in the
order you can find them on the playlist page

I used the date of the first video as the series' date. If that works,
again, the script could use that date in the _index.md of the series.

These points that I brought up are probably already obvious to you since
you've used this script before. I'm not an expert on Python and I'm not
even sure if it's possible to do what I mentioned (even though I doubt
it wouldn't be). If you want, I can get started working on it, if not,
I'll add the other series :)