Youtube playlist converter script improvements added
It's now possible to use options when calling the script through the
CLI. Use --help to find the full list with descriptions.

The previous behavior of the script is still somewhat present. Just call
the script like so:
```bash
$ python convert_youube_playlists_to_webpages
<PLAYLIST_ID>
```
and it will create a single folder at the root of the
script's project with the new files.

To create a playlist in the correct folder inside the website project,
use:
```bash
$ python convert_youtube_playlists_to_webpages.py <PLAYLIST_ID> --path "C:\!GDQuest\Repos\forks\GDquest-website\" --folders "game-design" "godot"
```

This will create a folder with the name of the playlist inside of `content/tutorial/game-design/godot`

All of what's left then is to fill in the information inside the indexes
of both the chapter and the series.

Closes #65