Fixed course section layout not generating when seconds is equal to zero
In the ocasion that there is a video with an exact number of minutes,
for instance 7:00, the layout would try to parse an empty string to int
and this would throw an error. A check was added to see if the string is
empty or not and only parse it in the case that it's not.