Fix getting started with Godot in 2021 youtube link
I have noticed that the getting started link redirects to a power sequencer tutorial. I do not know what a power seqeuncer is, but it seems incorrect. I searched on the GDQuest channel, and it seams this is the correct link.

I have never worked with this md file to build pages, and the fix is based on the [Making Your First Game: Basics] link reference, I could not test it correctly.